Create AddItemContext that will handle the addition of products
This AddItemContext was created so that the product addition functionality can be extended to the product list container.
